requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-jazzmin==2.6.0
django-smart-selects==1.6.0
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pypng==0.20220715.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
